feat: filter RPC methods from daemon
